x64 build added for windows; size_t re-factoring... updater now uses MD5 hashes instead of date and time; could be very unstable, though genrelly seems to work fine; moved some obsolete files away... next is to check if linux is still happy

+++ trunk/glib/gdrawobject.cpp
@@ -32,7 +32,7 @@ AGDrawObject::AGDrawObject(TGlRenderer& parent, const olxstr& collectionName) :
 AGDrawObject::~AGDrawObject()  {  ;  }
 //..............................................................................
 void AGDrawObject::Compile()  {
-  for( int i=0; i < Primitives->PrimitiveCount(); i++ )
+  for( size_t i=0; i < Primitives->PrimitiveCount(); i++ )
     Primitives->GetPrimitive(i).Compile();
 }
 //..............................................................................
