+++ trunk/repository/patchapi.h
@@ -56,7 +56,7 @@ public:
   static const char* GetUpdaterCmdFileName()  {  return "__cmds.update";  }
   static const char* GetOlex2PIDFileExt()  {  return "olex2_pid";  }
   static bool IsOlex2Running() {  return GetNumberOfOlex2Running() != 0;  }
-  static int GetNumberOfOlex2Running();
+  static size_t GetNumberOfOlex2Running();
   static bool LockUpdater();
   static bool UnlockUpdater();
